Moved Drawer navigator to separate file.
